fix / simplify parsing of lambda arguments
lambda arguments can be parsed without using try at all:

First, we decide if we have a set or an identifier. If we have a set,
there are two cases:

  * the lambda argument is a plain set (not followed by an `@` character)
  * or the set is optionally followed by `@<identifier>`, in which
    case the argument is a <right at> pattern.

If we got an identifier, we can do a similar thing, there are also two
cases:

  * plain identifier: not followed by `@`.
  * the idenfitifier is followed by `@<set>`. In this case, the argument
    is a <left at> pattern.

Avoiding excess use of try should improve performance and lead to better
error messages.