Make thunks monadic values
Before, we just enforced that with the `pure $!` during elimination. Now
we store them as a monadic value with `pure $!` or similar effectively
done already. Effectively, instead of encapsulation on the construction
and elimination side, we only have encapsulation on the construction
side.

In both the old and the new cases, the invariant is that forcing a thunk
must be done in the Monad needed to run the thunk lest it not be cached.

This shouldn't change the correctness of the existing code, at least as
measured by the test suite. But hopefully it makes writing code going
forward easier.