README: Getting Started: upd use new-* cabal options for Nix
Otherwise:
```
cabal test
Warning: The test command is a part of the legacy v1 style of cabal usage.

Please switch to using either the new project style and the new-test command
or the legacy v1-test alias as new-style projects will become the default in
the next version of cabal-install. Please file a bug if you cannot replicate a
working v1- use case with the new-style commands.

For more information, see: https://wiki.haskell.org/Cabal/NewBuild

/nix/store/iajira937isi5mjnrdcnvazd6p63b31d-ghc-8.6.3-with-packages/bin/ghc-pkg: createProcess: runInteractiveProcess: exec: does not exist (No such file or directory)
```

So at the very least `cabal new-test` seems to be required.