CI: Nixpkgs-Linux-main: make all builds optional
Due to Nixpkgs brittleness - it often blocks the development process, when the
PRs do not pass due to issues of downstream Nixpkgs.

Moreover that HNix builds in Nixpkgs with Nix, so the point of this builds is
just to check that HNix builds with Nixpkgs under Nix.

When Nixpkgs build is mandatory - PRs get blocked by it and PRs await on fixing
Nixpkgs.

Making Nixpkgs builds optional allows parallelization of code development and
Nixpkgs fixing.

More in discussion: https://github.com/haskell-nix/hnix/pull/681#discussion_r463983376