CI: GitHub: Nix: make GHC 8.8 nonstrict (#782)
For quite some time 8.8 strict build fails:
```
src/Nix/Expr/Types.hs:179:10: error: [-Wmissing-methods, -Werror=missing-methods]
    • No explicit implementation for
        ‘liftTyped’
    • In the instance declaration for ‘Lift (Fix NExprF)’
    |
179 | instance Lift (Fix NExprF) where
    |          ^^^^^^^^^^^^^^^^^
```

`nomissing-methods` seems like a bad path in this case and module.

I tried to address the issue and so far not constructed the implementation.

Too bad that the issue creeped-up backwards. Strict builds worked until this one
cropped-in.

So lowering the CI strictness until this gets resolved. Opened #783 #784 for that.