cabal: rm cryptohash-sha512 override (#814)
https://github.com/haskell-hvr/cryptohash-sha512/pull/5#issuecomment-757068529

Nixpkgs is Ok with all this, it does not import these overrides and have own separate mechanics for the overrides. Since cryptohash-sha512 package is essentially in the unmaintained state - I think it is not worth cleaning the Nixpkgs description because soon (on soon arriving 9.0) - it may be needed again in Nixpgs store, and to that time I hope we make a transition.

This cleaning has a nice side-effect that after release HNix would not need to recompile the cryptohash and Store packages every build.