repl: use pushScope to pass results from --eval -E
Improve my slopy `input` prototype from f7704b4.

```
hnix --repl --eval -E '{ a = 42; }'
hnix> a
42
```

In a case when argument passed to `-E` is not an attribute set,
we can't push it to scope directly, so we wrap it
in attribute set under `input` attribute, like previous version.

```
hnix --repl --eval -E '42'
hnix> input
42
```

Related to #292 #172.