treewide: unflip the `class MonadValue` `demand` & all its implementation
Argument order change has functional argument - so it is not possible to
elegantly switch into new code, and requires to go though transition.

So doing the work while doing needed refactoring at the same time.

This style of code currenly may seem more noizy, but really it is more
straight-forward, it mentions only operations & transformatios, types can be
looked in the HLS.

With the future work https://github.com/haskell-nix/hnix/issues/850
this style of the code would radically start to simplify itself, so please bear
with me.