Utils: inline what is definitely inlinable
Code is literally the code from the `base: Data.Bool`, there it is also does not
have inline directive:
```haskell
-- | From @Data.Bool ( bool )@.
bool :: a -> a -> Bool -> a
bool f _ False = f
bool _ t True  = t
```
Through profiler noticed that it was drawing resources.

GHC was not inlining the `bool`. Imagine.

So starting giving the nodges to the GHC.

This Utils funcitons are definit start.