Thunk.Basic: forceEff: refactor; enabling `ApplicativeDo`
Notice, here in `forceEff` I moved the thunk locking closer to where it is
needed, so function does a couple of preparation computations before locking the
thunk to operate on it.

Also notice that the `force` add `forceEff` look like 1 function, and as
`forceEff` also probably needs to check for exception when doing an action -
they are literally the same function.