String.Coerce: refactor pattern match
Most of it is a binary choice, made the logic of it seen better.

The structure of the code overall seems more understandable

Result of the refactoring is that code no longer pattern matches `CoerceStringy` on `CoerceAny`
cases.

Added a note that URI is not matched, which is strange, but currently do not
know what to think of it. NURI is Text.