Builtins: clean-up names, form `*Nix` group for builtins
Noticed that the `*_` was used to avoid collision with well-known Haskell
function names.

But `*_` has a different semantic meaning in Haskell, and the reader still thought it
is Haskell well-known functions with `_` modification.

This also created a dichotomy in function names, now some were normal, and some
were `_` without difference in the semantic meaning of them.

They now are formed in subgroup. All of them are internal module functions.

This subgroup can be put in a separate module and imported qualified, and so -
would be able to drop the `*Nix` suffix from them.