Builtins: seqNix: (const (pure b) <=< -> b <$)
This is just an eqivalent replacer. The reference implementation sequentially does for `a` &
`b` forceValue, and returns `b`.

The `seq` in GHC promises that `a` would be evaluated to `WHNF` - not more & not
less.

In this current implementation is `m` preservation from `a` is a good things,
idk so far.

The first version of `seq` was `a >> b`.

Then it became this one.

Since reference does forceValue bor both, the exact same implementation is:

```haskell
seqNix a b =
  demand b <* demand a
```