cabal: curb -O levels
The `-03` does not exist (but in plans) & (currently) anything >2 is a synonym
for 2.

Assumed C-lang-kind options & there is `-optc-O3` for them in GHC.

Also tried to move this switch, as Cabal team advises - to `cabal.project`.
But seems like `cabal.project` markup/parsing does not allow that, not found a
way to include `ghc-options` so they parse & work.