{Effects.Derivation, Reduce, Standard}: HashMap import clean-up
The is only 1 implementation of HashMap - with strict constructors, so
singleton also, `insert` is strict, `fromList` of course takes list lazily.

All those have 1 implementation adhearing to logic, and so `HashMap.Lazy` &
`HashMap.Strict` are equal in that regard and reexport the same implementations from
`HashMap.Internal`, so there is no need to declare HashMap.Strict or Lazy, and
not need to distinquish from where aforementioned functins got imported, so no
need to qualify them if that is not required.