treewide: use `one` for all singletons
Ok. I cave in on it in the end, it is a great idea.
`singleton` set of functions requires proper binding ad-hoc polymorphism.

And an-hoc polymorphism (TypeFamilies includded) is a native lambda space to Haskell.

The only thing that is needed is to check during writing is that type gets
inferred & no in the immediate recursion loop.