Cited.Basic: Cited: rm singleton field name
I strongly consider reation of records for singletons an antipattern.

Why to create some unique identificator to a single field, if there is only 1 field.

Moreover `getCited` name does not give any semantic information of from where & what it returns.

Newtype mechanism is an isomorphism, so one could ask "why someone creates a
custom functions to unwrap a newtype isomorphism". When for example, the
`coerse` exists, and since GHC guarantees equality of runtime representation for
newtype, that means that GHC guaranteese newtype unwrapping to be free, which
means guatantee of a free `coerse`.

Often I check & find that sigleton field accessors are never used, which is also
true in this case.

M  src/Nix/Cited/Basic.hs