Fix some issues revealed by static analysis (#960)
I had a run with [stan](https://github.com/kowainik/stan) an I have uploaded the [stan.html](https://maudoux.be/stan_original.html) report.

The tool is a bit heavy on strict data and avoiding unnamed black holes in pattern matches. But at least it made me remove some uses of Unsafe.

Another useful but weird case is that it highlights issues with the list type which theoretically allows infinite lists. Functions like length may thus not terminate. In our case moving to slists is not a good idea, but Vectors may represent better the lists we handle. But it is not worth switching if we create vectors from lists.