treewide: (Expr.Types.VarName -> newtype) lift-up type boundary
This is obviously a big change.

Before this units of data in expressions were undistinquisable from Text.

Not VarName is there to mark something as an abstraction, some argument that
has/can have/can be binded a value.

Went through the whole project & established the boundary.

The boundary stops entering the Derivations (their file format puts everything
in "", so they are left as text), CLI Options & Executable.

Maybe `KeyMap` datatype that is in utils should be removed, at least the
`Keymap SourcePos` should be replaced with `PosSet`.

Now the code would be more intuitive & would read better.