Expr: Types: Binding(Inherit): NKeyName -> VarName
Thread: https://github.com/haskell-nix/hnix/issues/377#issuecomment-883380109

`inherit x y` in y` position always takes a variable name.

Nix allows `inherit x "y"`, but there is no use (in the wild real life use) for
it, it seems a misfeature and would be considered a quirk of the original type
system/implementation, until the use case of it would be clear (which is hard,
since there is a single use of it in Nixpkgs, which is mentioned in the thread).