treewide: mv ViewPatterns into default extentions
Due to Haskell having coerce, and SPJ saying - it is the only one big change to
the core Haskell language they had like ever. That confirms my understanding of
`coerce` as the central feature of Haskell.

To use `coerce` comfortable & succinctly, the ViewPatterns use is needed,
therefore making it default extention.